Move LocationFilterMultiAutocomplete from artifact to character
