+++ libs/gi/ui/src/components/artifact/index.ts
@@ -10,7 +10,6 @@ export * from './ArtifactSetTooltip'
 export * from './ArtifactStatKeyDisplay'
 export * from './ArtifactTrans'
 export * from './IconStatDisplay'
-export * from './LocationFilterMultiAutocomplete'
 export * from './SetEffectDisplay'
 export * from './SetInclusionButton'
 export * from './SmolProgress'
