+++ libs/gi/ui/src/components/character/index.ts
@@ -9,6 +9,7 @@ export * from './CharacterSort'
 export * from './ConstellationDropdown'
 export * from './EquippedGrid'
 export * from './LocationAutocomplete'
+export * from './LocationFilterMultiAutocomplete'
 export * from './LocationName'
 export * from './StatDisplayComponent'
 export * from './TalentDropdown'
